Conflicts:
	src/ServiceStack.ServiceInterface/Testing/TestBase.cs
	src/ServiceStack/ServiceHost/ServiceController.cs
	src/ServiceStack/WebHost.Endpoints/Extensions/IHttpResponseExtensions.cs
	src/ServiceStack/WebHost.Endpoints/Support/EndpointHandlerBase.cs
	tests/ServiceStack.WebHost.Endpoints.Tests/Support/EndpointHandlerBaseTests.cs

+++ src/ServiceStack/WebHost.EndPoints/Handlers/AsyncHandlerBase.cs
@@ -59,7 +59,7 @@ namespace ServiceStack.WebHost.Endpoints.Handlers
 				//if there is a problem writing to response, by now it will be closed
 				if (!httpRes.IsClosed)
 				{
-                    httpRes.WriteErrorToResponse(httpReq.ResponseContentType, httpReq.OperationName, errorMessage, ex, statusCode);
+                    httpRes.WriteErrorToResponse(httpReq.ResponseContentType, httpReq.OperationName, errorMessage, ex, (int)statusCode);
 				}
 			}
 			catch (Exception writeErrorEx)
